Rename globalflags package
